Added support for versioning in the portfolio items (and removed them in the portfolio's, subportfolio's)

+++ dokeos/application/lib/portfolio/data_manager/database.class.php
@@ -68,33 +68,36 @@ class DatabasePortfolioDataManager extends PortfolioDataManager
 		return $succes;
 	}
 
-	function update_portfolio_publication($portfolio_publication)
+	function update_portfolio_publication($portfolio_publication, $delete_targets = true)
 	{
 		$condition = new EqualityCondition(PortfolioPublication :: PROPERTY_ID, $portfolio_publication->get_id());
 		$succes = $this->database->update($portfolio_publication, $condition);
 
-		$condition = new EqualityCondition(PortfolioPublicationGroup :: PROPERTY_PORTFOLIO_PUBLICATION, $portfolio_publication->get_id());
-		$succes &= $this->database->delete(PortfolioPublicationGroup :: get_table_name(), $condition);
-
-		$condition = new EqualityCondition(PortfolioPublicationUser :: PROPERTY_PORTFOLIO_PUBLICATION, $portfolio_publication->get_id());
-		$succes &= $this->database->delete(PortfolioPublicationUser :: get_table_name(), $condition);
-
-		foreach($portfolio_publication->get_target_groups() as $group)
+		if($delete_targets)
 		{
-			$pfpg = new PortfolioPublicationGroup();
-			$pfpg->set_portfolio_publication($portfolio_publication->get_id());
-			$pfpg->set_group_id($group);
-			$succes &= $pfpg->create();
+			$condition = new EqualityCondition(PortfolioPublicationGroup :: PROPERTY_PORTFOLIO_PUBLICATION, $portfolio_publication->get_id());
+			$succes &= $this->database->delete(PortfolioPublicationGroup :: get_table_name(), $condition);
+	
+			$condition = new EqualityCondition(PortfolioPublicationUser :: PROPERTY_PORTFOLIO_PUBLICATION, $portfolio_publication->get_id());
+			$succes &= $this->database->delete(PortfolioPublicationUser :: get_table_name(), $condition);
+	
+			foreach($portfolio_publication->get_target_groups() as $group)
+			{
+				$pfpg = new PortfolioPublicationGroup();
+				$pfpg->set_portfolio_publication($portfolio_publication->get_id());
+				$pfpg->set_group_id($group);
+				$succes &= $pfpg->create();
+			}
+	
+			foreach($portfolio_publication->get_target_users() as $user)
+			{
+				$pfpu = new PortfolioPublicationUser();
+				$pfpu->set_portfolio_publication($portfolio_publication->get_id());
+				$pfpu->set_user($user);
+				$succes &= $pfpu->create();
+			}
 		}
-
-		foreach($portfolio_publication->get_target_users() as $user)
-		{
-			$pfpu = new PortfolioPublicationUser();
-			$pfpu->set_portfolio_publication($portfolio_publication->get_id());
-			$pfpu->set_user($user);
-			$succes &= $pfpu->create();
-		}
-
+		
 		return $succes;
 	}
